requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
streamlit
gunicorn

# Gemini
google-generativeai
